// EcomWeb/src/api/client/models/RegisterRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisterRequest = {
    name?: string | null;
    email?: string | null;
    phoneNumber?: string | null;
    dob?: string;
    password?: string | null;
    confirmPassword?: string | null;
};